zus: force dependency between modules and libzus
In order to support proper parallel compilation we need libzus to be
available before compiling the libfs modules, otherwise they fail during
linkage.

Had to move the include in the Makefile in order for fs/Makefile to see
the LIBZUS variable, which is a bit weird as make should resolve these
after the include, but whatever.

Signed-off-by: Yigal Korman <yigalk@netapp.com>