zus: enable coredumps for zusd (PXS-3195)
Enable coredumps in systemd zusd service.

In CentOS/RHEL the default core_pattern (how and where the dump is
created) is just 'core' so in order to redirect zusd core dumps
somewhere specific, we configure the working directory to point to
/var/log/zufs. The coredumps will be created there.

Also limit the size of the core dump. 50M is arbitrary, it big enough
to contain everything but not too big so we can collect it with our
support script.

Signed-off-by: Yigal Korman <yigalk@netapp.com>