zus: resilient zusd service recovery (PXS-3194)
We want to recover in case of zusd crash.
We use the systemd "Restart" feature to recover.

For now we don't support recovery of mounted libfs after zusd crash
(it's planned).
So we must have a proper shutdown of zufs libfs and rootfs before
loading zusd back, this is what the zusd.down script does -
tries to kill all processes locking the mount point and then umount.

Signed-off-by: Yigal Korman <yigalk@netapp.com>