pkg: fix issues in packaging around zufs.conf (PXS-3649)
It might be removed by the admin which is fine, so use 'rm -f' to ignore
errors.

And using $DESTDIR in uninstall didn't make sense as we don't use it
outside of the packaging logic.

V2:
Didn't read the Makefile properly, only executes post_install if DESTDIR
is not configured, so no reason to use it in the script.
I think we had a failed install on a host with the env var DESTDIR
configured to something.

V3:
Upgrade scenario was broken - can't call "return" from something that isn't
a function, better copy the behaviour from other scripts we have.

Signed-off-by: Yigal Korman <yigalk@netapp.com>
Signed-off-by: Nave Vardy <nave@netapp.com>