toyfs: Sample pmem-base file-system via ZUFS
ToyFS is nothing but a demo file system via zufs with ephemeral
semantice. It uses a pmam or volatile ram upon mount and invalidates
it upon umount. Directory listing and regular-file mapping are done
using simple double linked-lists, so performance is sub-sub optimal.

The main purpose of ToyFS is educational: demonstrate how to build
user-space file-system via ZUFS. Do *not* try to use it in production.

Signed-off-by: Shachar Sharon <sshachar@netapp.com>