build: revise moudule types handling
As the GENERIC and STANDALONE names were a bit confusing, change the
respectivle to ZUS_BIN and ZUS_NONE where filesystems will be ZUS_FS. As
the names suggest:

ZUS_FS is a zus file system
ZUS_BIN is a binary that is linked to the zus library and isn't a file
system
ZUS_NONE is a binary that uses the ZUS build system, but isn't tightly
linked to the zus echo system.

This patch also adds a check at the beginning of the build to see that
any calls for a module build is one those three.

Signed-off-by: Omer Caspi <omer.caspi@netapp.com>