build: common target defs in common makefile
Move generic build definitions regarding the target type (dynamic
library or executable) and it's naming into the common makefile. This
simplifies the different client makefiles (ZUS` and others), as correct
assignment of the flags and target naming is done in one place.

Also, adds verification that a project has a name, and its target type
is valid (i.e. lib or bin).

Signed-off-by: Omer Caspi <omer.caspi@netapp.com>