zus: fix huge number of vma splits (vm.max_map_count) caused by pa
vma split happens when managing pages with different flags,
for example locked vs non-locked.
on huge systems we hit a sysctl limit of 65530 in vm.max_map_count.
The solution is:
do not munlock on __pa_free in order to prevent (not needed) split,
because the page is going to be deallocated by fba_punch_hole anyway.
On the other hand when we pa_alloc_order some page(s)
we populate them with mlock, without pagefault, into the same (locked) vma.

Signed-off-by: Vladislav Drouker <vladislav.drouker@netapp.com>
Signed-off-by: Sagi Manole <sagim@netapp.com>