pkg: fix package install on unsupported kernels
If we don't have a zuf module for the running kernel, we will try to
create a bad symlink and modprobe will fail.
Instead, check properly that the module exists.

The bash trickery in the 'ln -s ...' line is in case we match more than
one kernel module (shouldn't happen, but just in case) - so we symlink
to the first one in the list.

Also, don't fail package installation if the service failed, let the
admin sort that out later (can install supported kernel if needed).

V2:
Same for uninstall.

Signed-off-by: Yigal Korman <yigalk@netapp.com>