zus: Use shared zus_page(s) from Kernel
pa_page must be the same as zus_page. Make sure its so.
We needed defined BUILD_BUG_ON define it.

Check that all members and sizes of pa_page align with
zus_page members

mmap to the Kernel's pages instead of using an fba allocation.

definition of fba was removed from md.h because it is no longer
used. Move it back to zus.h where it belongs

[ver 2]
  map size must align on 2M and never less so Kernel will give us
  a 2M aligned pointer. Some famous Netapp FS has a BUG if this is
  not so.

Signed-off-by: Boaz Harrosh <boazh@netapp.com>